requirements.txt
requests
matrix-nio
chess
